requirements.txt
# Voice to Text Mac — Dependencies
# Python 3.11+ required

# Global hotkey listener
pynput>=1.7.6

# Audio recording
sounddevice>=0.5.0
scipy>=1.10.0

# Clipboard operations
pyperclip>=1.8.2

# OpenAI Whisper API client
openai>=1.52.0

# Environment variables
python-dotenv>=1.0.1

# macOS menu bar
rumps>=0.4.0

# Image support
Pillow>=10.4.0

# macOS accessibility (optional, for better key simulation)
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
